fixed delete button; added link to home in header
