+++ client/src/components/Nav/index.css
@@ -1,0 +1,15 @@
+.navbar {
+  background-color: var(--light-grey);
+}
+
+.navbar ul {
+  border: 1px solid greenyellow;
+}
+
+.navbar li {
+  display: inline;
+  color: white;
+}
+
+.navbar-link {
+}
