+++ client/src/components/Post/index.js
@@ -1,24 +1,16 @@
 import './index.css';
-import { useMutation } from '@apollo/client';
-import { REMOVE_POST } from '../../utils/mutations';
 
 const Post = ({ postDetails }, { loggedInArtist }) => {
-  const [removePost] = useMutation(REMOVE_POST);
-
   // console.log(postDetails);
   const { _id, title, image, description, createdBy } = postDetails;
 
-  const handleDeleteClick = async (event) => {
-    const postId = _id;
-
-    const deletePost = await removePost({
-      variables: {
-        postId,
-        createdBy: loggedInArtist
-      }
-    });
-    console.log('deletedPost: ', deletePost);
-  };
+  // let isPostCreator = false;
+  // if (loggedInArtist === createdBy) {
+  //   isPostCreator = true;
+  // }
+  // console.log('logged in artist: ', loggedInArtist);
+  // console.log('createdBy: ', createdBy);
+  // console.log('isPostCreator: ', isPostCreator);
 
   return (
     <div className="post">
@@ -26,11 +18,6 @@ const Post = ({ postDetails }, { loggedInArtist }) => {
       <img src={`data:image/png;base64,${image}`} alt={description} />
       <p>{description}</p>
       <p>created by: {createdBy}</p>
-      <div>
-        <button type="button" onClick={handleDeleteClick}>
-          Delete Post
-        </button>
-      </div>
     </div>
   );
 };
